% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLR_Utils.R
\name{calibrated_classifier}
\alias{calibrated_classifier}
\title{Train a calibrated classifier using platt or isotonic regression}
\usage{
calibrated_classifier(y_true, y_pred, method = "isotonic")
}
\arguments{
\item{y_true}{Vector of true labels. Should be binary (0 or 1)}

\item{y_pred}{Vector of predicted probabilities from a classifier}

\item{method}{Character, either "isotonic" (default) or "platt"}
}
\value{
A calibrated_classifier S3 class object
}
\description{
Train a calibrated classifier using platt or isotonic regression
}
