% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibratedClassifier.R
\name{calibration_curve}
\alias{calibration_curve}
\title{Plots a reliability chart of a set of predicitons from a classifier}
\usage{
calibration_curve(y_true, y_pred, bins = 10, scale = T)
}
\arguments{
\item{y_true}{Vector of true labels. Should be binary (0 or 1)}

\item{y_pred}{Vector of predictions from the classifier
Should be real number.}

\item{bins}{The number of bins to use in the reliability plot}

\item{scale}{Scale the pred to be between 0 and 1 before creating reliability plot}
}
\value{
data.frame of mean prediction value vs fraction of positives
}
\description{
Plots a reliability chart of a set of predicitons from a classifier
}
