% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLR_Utils.R
\name{fit_isoreg}
\alias{fit_isoreg}
\title{Fit isotonic regression}
\usage{
fit_isoreg(iso, x0)
}
\arguments{
\item{iso}{model from the stats::isoreg function}

\item{x0}{vector of probabilities from a classifier}
}
\value{
A fitted isotonic regression model
}
\description{
Fit isotonic regression
}
