% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyPreprocWrapper.R
\name{makePreprocWrapperDummyFeatures}
\alias{makePreprocWrapperDummyFeatures}
\title{Preprocessing wrapper for one-hot encoding}
\usage{
makePreprocWrapperDummyFeatures(learner, method = "1-of-n")
}
\arguments{
\item{learner}{mlr learner}

\item{method}{character, one of c("1-of-n", "reference")}
}
\value{
PreprocWrapper
}
\description{
Preprocessing wrapper for one-hot encoding
}
