% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalePreprocWrapper.R
\name{makePreprocWrapperScale}
\alias{makePreprocWrapperScale}
\title{Preprocessing scaling wrapper}
\usage{
makePreprocWrapperScale(learner, center_data = TRUE, scale_data = TRUE)
}
\arguments{
\item{learner}{mlr learner}

\item{center_data}{logical, default = TRUE}

\item{scale_data}{logical, default = TRUE}
}
\value{
PreprocWrapper
}
\description{
Preprocessing scaling wrapper
}
