% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrUtils.R
\docType{package}
\name{mlrUtils}
\alias{mlrUtils}
\alias{mlrUtils-package}
\title{mlrUtils: Miscillaneous functions to be used with the `mlr` package}
\description{
mlrUtils: Miscillaneous functions to be used with the `mlr` package
}
\author{
Steven Pawley, \email{dr.stevenpawley@gmail.com}
}
