% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationResamplingImportance.R
\name{permutationResamplingImportance}
\alias{permutationResamplingImportance}
\title{Permutation importances calculated during resampling}
\usage{
permutationResamplingImportance(learner, task, resampling, measure,
  n_permutations = 50)
}
\arguments{
\item{learner}{Learner, WrappedLearner}

\item{task}{task}

\item{resampling}{ResampleDesc object}

\item{measure}{mlr performance measure}

\item{n_permutations}{numeric, number of permutations, default = 50}
}
\value{
dataframe, containing permutations scores per feature
}
\description{
Permutation importances calculated during resampling
}
