% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predFun.R
\name{predFun}
\alias{predFun}
\title{Prediction function for mlr regression model applied to raster data}
\usage{
predFun(model, data, lev = NULL)
}
\arguments{
\item{model}{mlr fitted model}

\item{data}{data passed from raster}

\item{lev}{named list of RasterLayer factor attribute tables. Each table
is a data.frame; the first column has to be named `ID` and contains a
list of unique pixel values in the raster; the second column contains the
factors that correspond to each pixel value.}
}
\value{
RasterLayer object
}
\description{
Prediction function for mlr regression model applied to raster data
}
