% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLR_Utils.R
\name{predfun_calibrated}
\alias{predfun_calibrated}
\title{Prediction function for raster::predict method to produce classification,
probabilities for a binary class, and optionally calibrated probabilities}
\usage{
predfun_calibrated(model, data, calmodel = NULL, type = "all",
  label = "prob.1")
}
\arguments{
\item{model}{fitted mlr model}

\item{data}{data passed from the raster predict method}

\item{calmodel}{calibrated_classifier model}

\item{type}{predict 'classification', 'probabilities', or 'all'}

\item{label}{name of class to produce probabilities for}
}
\value{
numeric or data.frame
}
\description{
Prediction function for raster::predict method to produce classification,
probabilities for a binary class, and optionally calibrated probabilities
}
