% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibratedClassifier.R
\name{predict.calibrated_classifier}
\alias{predict.calibrated_classifier}
\title{Predict method for the calibrated_classifier S3 class}
\usage{
\method{predict}{calibrated_classifier}(model, y_pred)
}
\arguments{
\item{model}{A calibrated_classifier S3 class}

\item{y_pred}{Vector of probabilities upon which to apply calibration.
The y_pred values should be disjoint from the data used to train the
calibrated_classifier model}
}
\value{
Vector of calibrated probabilities.
}
\description{
Predict method for the calibrated_classifier S3 class
}
