% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLR_Utils.R
\name{predictStandardRescaler}
\alias{predictStandardRescaler}
\title{Prediction function for custom preprocessor wrapper to center and scale data}
\usage{
predictStandardRescaler(data, target, args, control)
}
\arguments{
\item{data}{data.frame. Newdata to predict}

\item{target}{character. Name of target (response) variable}

\item{args}{list. Arguments to center and/or scale the data}

\item{control}{scaling parameters from trainfun to be used in the prediction}
}
\description{
Prediction function that is required for custom preprocessor wrapper in order
to apply the wrapper to new data.
}
