% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLR_Utils.R
\name{trainStandardRescaler}
\alias{trainStandardRescaler}
\title{Training function for custom preprocessor wrapper to center and scale data}
\usage{
trainStandardRescaler(data, target, args = list(center = T, scale = T))
}
\arguments{
\item{data}{data.frame. Training data including response variable}

\item{target}{character. Name of target (response) variable}

\item{args}{list. Arguments to center and/or scale the data}
}
\description{
Training function for custom preprocessor wrapper to center and scale data
}
